feat: add quick-login function
If quicklogin-enabled== true in the seat item in /etc/lightdm/lightdm.conf is true,
it is a quick login situation, and the automatic login process will be used.
First determine whether there is a record of the last login user and if it is in
the automatic login user list, then automatically log in the last user.
/var/lib/lightdm/lightdm-deepin-greeter/state_user.
If there is no record or the configuration file does not exist,
the original process will be used. And DDE_QUICKLOGIN == true
is added to the environment variable.
If automatic login is enabled, the automatic login process will be used,
ignoring the quick login configuration.

Log: as title
Pms: TASK-377223